add spock functions to non-relic version
This also refactor the bls errors into a non-tag specific file

This is needed to unblock #3696 (SPOCKProve is now used by the emulator
intergration testing)

gh: #3695
