+++ crypto/bls_no_relic_test.go
@@ -20,9 +20,6 @@ func TestNoRelicPanic(t *testing.T) {
 	assert.PanicsWithValue(t, relic_panic, func() { AggregateBLSPrivateKeys(nil) })
 	assert.PanicsWithValue(t, relic_panic, func() { AggregateBLSPublicKeys(nil) })
 	assert.PanicsWithValue(t, relic_panic, func() { IdentityBLSPublicKey() })
-	assert.PanicsWithValue(t, relic_panic, func() { IsBLSAggregateEmptyListError(nil) })
-	assert.PanicsWithValue(t, relic_panic, func() { IsInvalidSignatureError(nil) })
-	assert.PanicsWithValue(t, relic_panic, func() { IsNotBLSKeyError(nil) })
 	assert.PanicsWithValue(t, relic_panic, func() { IsBLSSignatureIdentity(nil) })
 	assert.PanicsWithValue(t, relic_panic, func() { RemoveBLSPublicKeys(nil, nil) })
 	assert.PanicsWithValue(t, relic_panic, func() { VerifyBLSSignatureOneMessage(nil, nil, nil, nil) })
